Update docs and examples
Updated documentation and example to highlight the new universal tornado
scope manager that works with all versions.

This requires users to import the TornadoScopeManager from
`tornado_opentracing.scope_managers` instead of
`opentracing.scope_managers.tornado`. This will be mentioned in detail
in the release docs.

We'll also submit a PR to python-opentracing and patch it to export the
correct scope manager based on Tornado and Python version being used so
the above change will be unnecesary.